backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
openai==1.52.2
python-jose==3.3.0
authlib==1.3.2
email-validator==2.2.0
orjson==3.10.7
